Release NEFTranslator v2.0.0, which enables bidirectional translation between NEF and NMR-STAR

+++ wwpdb/utils/tests-nmr/make_clean_NmrDpUtility.py
@@ -64,13 +64,13 @@ class TestNmrDpUtility(unittest.TestCase):
         with open(self.data_dir_path + entry_id + '-clean-deposit-log.json', 'r') as file:
             report = json.loads(file.read())
 
-        if not report['error'] is None:
+        if not report['error'] is None and os.path.exists(self.data_dir_path + entry_id + '-clean.nef'):
             os.remove(self.data_dir_path + entry_id + '-clean.nef')
 
         with open(self.data_dir_path + entry_id + '-clean-str-deposit-log.json', 'r') as file:
             report = json.loads(file.read())
 
-        if not report['error'] is None:
+        if not report['error'] is None and os.path.exists(self.data_dir_path + entry_id + '-clean.str'):
             os.remove(self.data_dir_path + entry_id + '-clean.str')
 
     def test_nmr_nef2str_deposit_check_1nk2(self):
